chore: refactoring to versions helper
